Network Layer (Repository Pattern)
