--- app/src/main/java/one/njk/celestidesk/data/auth/AuthRepository.kt
@@ -1,11 +1,0 @@
-package one.njk.celestidesk.data.auth
-
-import one.njk.celestidesk.data.auth.model.AuthLoginRequest
-import one.njk.celestidesk.data.auth.model.AuthResult
-import one.njk.celestidesk.data.auth.model.AuthSignUpRequest
-
-interface AuthRepository {
-    suspend fun signUp(user: AuthSignUpRequest): AuthResult<Unit>
-    suspend fun logIn(user: AuthLoginRequest): AuthResult<Unit>
-    suspend fun authenticate(): AuthResult<Unit>
-}
