+++ app/src/main/java/one/njk/celestidesk/network/auth/AuthRepositoryImpl.kt
@@ -1,14 +1,15 @@
-package one.njk.celestidesk.data.auth
+package one.njk.celestidesk.network.auth
 
 import android.util.Log
 import one.njk.celestidesk.data.RolesDataStore
-import one.njk.celestidesk.data.auth.model.AuthLoginRequest
-import one.njk.celestidesk.data.auth.model.AuthResult
-import one.njk.celestidesk.data.auth.model.AuthSignUpRequest
+import one.njk.celestidesk.network.auth.model.AuthLoginRequest
+import one.njk.celestidesk.network.auth.model.AuthResult
+import one.njk.celestidesk.network.auth.model.AuthSignUpRequest
+import one.njk.celestidesk.network.ApiService
 import retrofit2.HttpException
 
 class AuthRepositoryImpl(
-    private val api: AuthApi,
+    private val api: ApiService,
     private val pref: RolesDataStore
 ): AuthRepository {
     override suspend fun signUp(user: AuthSignUpRequest): AuthResult<Unit> {
