+++ app/src/main/java/one/njk/celestidesk/network/auth/model/AuthLoginRequest.kt
@@ -1,4 +1,4 @@
-package one.njk.celestidesk.data.auth.model
+package one.njk.celestidesk.network.auth.model
 
 data class AuthLoginRequest(
     val username: String,
